Reorganize, no GUI, add corrections, add plotter
The code is organized into a set of functions instead of a script.
The nais_processer() handles the processing and the nais_plotter()
handles the plotting.

Instead of GUI the user input is handled by keyword arguments and
command line input.

Add a correction for changes in sample flow and correct the ion mode
based on a calibration by Wagner et al. (2016),
http://hdl.handle.net/10138/165313.